Add caching of contests, problems, news & profile A huge improvement in usability!
The data gets cached in text files in cachedir with different TTLs.
Problem contents will be an LRU cache with a size limit.
Results will probably not be cached as they oftentimes need to get
updated every minute or so and caching them with such a small TTL is
kind of pointless.
Next step would be making the TTLs configurable from a satori.conf file
in configdir and documenting all that.
Also changed get_login_credentials to take a keyword argument saying
whether the password is needed, and changed a couple of lines of code to
use isempty instead of comparing length to 0.
Updated todo and readme.